refactor(examples): import Cargo.toml (importing lib/main should be an anti-pattern)
